requirements.txt
requests==2.31.0
supabase==2.17.0
ddgs==9.5.1
python-dotenv==1.0.1
pillow==10.1.0
rich==13.7.0